Add gravacao usuario e lastsync

+++ quantisync/lib/util.py
@@ -1,4 +1,5 @@
 from pathlib import Path
+from datetime import datetime
 import glob
 
 
@@ -26,7 +27,8 @@ class File:
         return str(self._path)
 
     def unlink(self):
-        return self._path.unlink()
+        if self.exists():
+            return self._path.unlink()
 
 
 class Dir:
@@ -39,3 +41,53 @@ class Dir:
             return []
 
         return glob.glob('{}/*.{}'.format(str(self._path), extension))
+
+
+class InvalidDateString:
+    pass
+
+
+class Date:
+    def parseString(string):
+        if not string:
+            raise InvalidDateString()
+
+        return datetime.strptime(string, '%Y-%m-%d %H:%M:%S.%f')
+
+
+class DeltaTime:
+
+    @classmethod
+    def _handleSuffix(cls, singularSuffix, pluralSuffix, diff):
+        string = 'há {} '.format(diff)
+        if diff > 1:
+            string = string + pluralSuffix
+        else:
+            string = string + singularSuffix
+        return string
+
+    @classmethod
+    def format(cls, delta):
+        DAYS_IN_YEAR = 365
+        DAYS_IN_MONTH = 30
+
+        SECONDS_IN_MIN = 60
+        SECONDS_IN_HOUR = 3600
+
+        if delta.days >= DAYS_IN_YEAR:
+            deltaYears = int(delta.days / DAYS_IN_YEAR)
+            return cls._handleSuffix('ano', 'anos', deltaYears)
+        elif delta.days >= DAYS_IN_MONTH:
+            deltaMonths = int(delta.days / DAYS_IN_MONTH)
+            return cls._handleSuffix('mês', 'meses', deltaMonths)
+        elif delta.days >= 1:
+            return cls._handleSuffix('dia', 'dias', delta.days)
+        else:
+            if delta.seconds < SECONDS_IN_MIN:
+                return 'agora'
+            elif delta.seconds >= SECONDS_IN_MIN and delta.seconds < SECONDS_IN_HOUR:
+                deltaMins = int(delta.seconds / SECONDS_IN_MIN)
+                return cls._handleSuffix('minuto', 'minutos', deltaMins)
+            else:
+                deltaHours = int(delta.seconds / SECONDS_IN_HOUR)
+                return cls._handleSuffix('hora', 'horas', deltaHours)
